requirements.txt
# FastAPI y servidor ASGI
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# Database y ORM
sqlmodel>=0.0.16
psycopg2-binary>=2.9.7
asyncpg>=0.29.0

# Redis
redis>=5.0.1

# Validación y serialización
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Utilities
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.7

# HTTP client
httpx>=0.26.0

# Logging y monitoring
structlog>=23.2.0

# Environment variables
python-dotenv>=1.0.1

# Data processing
pandas>=2.2.0
numpy>=1.26.0
